vcta_django/requirements.txt
asgiref>=1.0,<1.0.99
channels>=1.1,<1.1.99
daphne>=1.0,<1.0.99
django-cors-headers>=2.0,<2.0.99
django>=1.10,<1.10.99
djangorestframework>=3.6,<3.6.99
ipython>=5.3,<5.3.99
redis>=2.10.5,<2.10.99
pypiwin32